requirements.txt
affine==2.3.0
amqp==2.5.2
asgiref==3.2.3
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
backcall==0.1.0
billiard==3.6.2.0
bleach==3.1.0
boto3==1.11.15
botocore==1.14.15
branca==0.3.1
cachetools==4.0.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.3.0
cogeotiff==0.1.4
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
daphne==2.4.1
dask==2.10.1
-e git+https://github.com/opendatacube/datacube-core.git@98cf9ba35d90bdc582dce6b399dc4ba274676bca#egg=datacube
decorator==4.4.1
defusedxml==0.6.0
Django==3.0
django-celery-beat==1.6.0
django-cors-headers==3.2.1
django-extensions==2.2.8
django-taggit==1.2.0
django-timezone-field==3.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
folium==0.10.1
GDAL==3.0.2
geojson==2.5.0
geomet==0.2.1.post1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
hdmedians==0.13
html2text==2020.1.16
hyperlink==19.0.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
kiwisolver==1.1.0
kombu==4.6.7
lark-parser==0.8.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
netCDF4==1.3.1
networkx==2.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.5.2
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyPEG2==2.15.2
pyrsistent==0.15.7
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
rasterio==1.0.24
redis==3.4.1
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.9
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.5.0
sentinelsat==0.13
service-identity==18.1.0
Shapely==1.7.0
singledispatch==3.4.0.3
six==1.14.0
snuggs==1.4.7
SQLAlchemy==1.3.1
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
xarray==0.15.0
zipp==2.2.0
zope.interface==4.7.1
